fixed routes, made admin and owner checks, and removed all the hardco…

+++ controllers/postController.js
@@ -17,8 +17,8 @@ const feed_list_get = async (req, res) => {
 };
 
 const feed_like = async (req, res) => {
-  console.log('like', req.params.id, req.params.user);
-  const like = await postModel.likePost(req.params.id, req.params.user);
+  console.log('like', req.params.id, req.user.userId);
+  const like = await postModel.likePost(req.params.id, req.user.userId);
   return res.json(like);
 }
 
@@ -28,12 +28,13 @@ const post_list_get_postedBy = async (req, res) => {
 };
 
 const post_add_comment = async (req, res) => {
-  console.log('add comment ', req.body.comment);
-  const comments = await postModel.addComment(req.params.postId, req.params.commenter, req.body.comment);
+  console.log('add comment ', req.body.comment, req.user.userId);
+  const comments = await postModel.addComment(req.params.postId, req.user.userId, req.body.comment);
   return res.json(comments);
 };
 
 const post_find_comments = async(req, res) => {
+  console.log(req.body, 'body ', req.user)
   const comments = await postModel.findComments(req.body);
   return res.json(comments);
 }
@@ -101,13 +102,28 @@ const post_add_ingredient = async (req, res) => {
 };
 
 const post_delete = async (req, res) => {
-  const deleteOk = await postModel.deletePost(req.params.id);
-  res.json(deleteOk);
+  try{
+    const owner = await postModel.getPost(req.params.id);
+    console.log('post_delete ', owner, " trying to delete")
+    if (req.user.userId === owner.userId || req.user.admin === 1){
+      const deleteOk = await postModel.deletePost(req.params.id);
+      res.json(deleteOk);
+    }
+  }catch(e){
+    res.status(403).json({error: e.message});
+  }
 };
 
 const post_update = async (req, res) => {
-  const deleteOk = await postModel.updatePost(req.params.id, req.body.caption);
-  res.json(deleteOk);
+  try{
+    const owner = await postModel.getPost(req.params.id);
+    if(req.user.userId === owner.userId){
+      const deleteOk = await postModel.updatePost(req.params.id, req.body.caption);
+      res.json(deleteOk);
+    }
+  }catch(e){
+    res.status(403).json({error: e.message});
+  }
 };
 
 const post_get_likes = async (req, res) => {
